refactor: move UpgradePrompt component to new location
Moves the UpgradePrompt component from the course-stage-step 
directory to the course-page directory. This change improves 
the organization of components by grouping related files 
together, enhancing maintainability and clarity in the codebase.
