+++ app/components/course-page/course-stage-step/inaccessible-community-solutions-list.hbs
@@ -15,7 +15,7 @@
   </:content>
   <:overlay>
     <div class="pt-4 px-4 sm:pt-8 sm:px-8 w-full">
-      <CoursePage::CourseStageStep::UpgradePrompt @featureSlugToHighlight="code-examples" class="w-full shadow-lg" />
+      <CoursePage::UpgradePrompt @featureSlugToHighlight="code-examples" class="w-full shadow-lg" />
     </div>
   </:overlay>
 </BlurredOverlay>
