+++ app/components/course-page/upgrade-prompt.hbs
@@ -11,7 +11,7 @@
       </div>
 
       <div class="prose dark:prose-invert prose-compact prose-sm @[640px]:prose-base py-3" data-test-upgrade-prompt-secondary-copy>
-        {{#if this.isLoadingRegionalDiscount}}
+        {{#if this.isLoadingDiscounts}}
           <p>
             Plans start at $30/month when billed annually.
           </p>
