+++ app/components/upgrade-modal.hbs
@@ -5,7 +5,7 @@
   {{did-insert this.handleDidInsert}}
   ...attributes
 >
-  <CoursePage::CourseStageStep::UpgradePrompt @featureSlugToHighlight={{@featureSlugToHighlight}} class="w-full" />
+  <CoursePage::UpgradePrompt @featureSlugToHighlight={{@featureSlugToHighlight}} class="w-full" />
 
   <div class="absolute right-0 top-0">
     <button type="button" class="w-12 h-12 flex items-center justify-center" {{on "click" @onClose}}>
