+++ app/templates/course/stage/instructions.hbs
@@ -5,7 +5,7 @@
 >
   <div class="pt-6 pb-32 px-3 md:px-6 lg:px-10" {{did-update this.handleDidUpdateTestsStatus this.currentStep.testsStatus}}>
     {{#if this.shouldShowUpgradePrompt}}
-      <CoursePage::CourseStageStep::UpgradePrompt @featureSlugToHighlight="content" class="mb-6" />
+      <CoursePage::UpgradePrompt @featureSlugToHighlight="content" class="mb-6" />
     {{/if}}
 
     {{#if this.shouldShowFeedbackPrompt}}
